# docs/openapi/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"LegacyRestrictionDetailsResponse": {
				"properties": {
					"response": {
						"properties": {
							"errorMsg": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							},
							"respMsg": {
								"type": "string"
							},
							"respCodeMsg": {
								"type": "string"
							},
							"respCode": {
								"type": "string"
							},
							"respStatus": {
								"type": "boolean"
							}
						},
						"required": [
							"errorMsg",
							"errorCode",
							"respMsg",
							"respCodeMsg",
							"respCode",
							"respStatus"
						],
						"type": "object"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"response",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyRestrictionDeniedResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"restrictions": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"result",
					"restrictions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyUserRestrictedRequest": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"customer",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFwcloudErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "FWCloud API",
		"version": "2.4.1",
		"description": "FWCloud RESTful API.",
		"contact": {}
	},
	"paths": {
...
		"/user/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"200": {
						"description": "Restriction details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyRestrictionDetailsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"response": {
												"respStatus": true,
												"respCode": "ACR_OK",
												"respCodeMsg": "Ok",
												"respMsg": "",
												"errorCode": "",
												"errorMsg": ""
											},
											"data": {}
										}
									}
								}
							}
						}
					},
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyRestrictionDeniedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check that there are no restrictions for user deletion.",
				"summary": "Restrictions for user deletion.",
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyUserRestrictedRequest"
							},
							"example": {
								"customer": 10,
								"user": 5
							}
						}
					}
				}
			}
		},
...
		"/fwcloud/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Check if the fwcloud indicated as a parameter has any deletion restriction.",
				"summary": "Check delete restrictions.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
...
		"/firewall/cloud/get": {
			"put": {
				"operationId": "GetByCloud",
				"responses": {
					"200": {
						"description": "Get firewalls in cloud",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFirewallResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"cluster": null,
												"fwcloud": 1,
												"name": "Firewall-01",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											},
											{
												"id": 2,
												"cluster": null,
												"fwcloud": 1,
												"name": "Firewall-02",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get firewalls data by fwcloud.",
				"summary": "Get firewalls in cloud.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallCloudRequest"
							},
							"example": {
								"fwcloud": 1
							}
						}
					}
				}
			}
		},
		"/firewall/cluster/get": {
			"put": {
				"operationId": "GetByCluster",
				"responses": {
					"200": {
						"description": "Get firewalls in cluster",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFirewallResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"cluster": 2,
												"fwcloud": 1,
												"name": "Firewall-01",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 1,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											},
											{
												"id": 2,
												"cluster": 2,
												"fwcloud": 1,
												"name": "Firewall-02",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get firewalls data by cluster.",
				"summary": "Get firewalls in cluster.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallClusterRequest"
							},
							"example": {
								"fwcloud": 1,
								"cluster": 2
							}
						}
					}
				}
			}
		},
...
		"/customer/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerRestrictionResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check that there are no restrictions for customer deletion.",
				"summary": "Restrictions for customer deletion.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerRequest"
							},
							"example": {
								"customer": 10
							}
						}
					}
				}
			}
		},
		"/cluster": {
			"post": {
				"operationId": "Store",
				"responses": {
					"200": {
						"description": "New cluster",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterStoreResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"insertId": 1,
											"loData": {}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Tree node access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new cluster of firewalls.",
				"summary": "New cluster.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterStoreRequest"
							},
							"example": {
								"fwcloud": 3,
								"node_id": 391,
								"clusterData": {
									"name": "Cluster-01",
									"options": 3,
									"fwnodes": [
										{
											"name": "node1",
											"comment": null,
											"install_user": null,
											"install_pass": null,
											"save_user_pass": 1,
											"install_interface": null,
											"install_ipobj": null,
											"fwmaster": 1,
											"install_port": 22
										},
										{
											"name": "node2",
											"comment": null,
											"install_user": null,
											"install_pass": null,
											"save_user_pass": 1,
											"install_interface": null,
											"install_ipobj": null,
											"fwmaster": 0,
											"install_port": 22
										},
										{
											"name": "node3",
											"comment": null,
											"install_user": null,
											"install_pass": null,
											"save_user_pass": 1,
											"install_interface": null,
											"install_ipobj": null,
											"fwmaster": 0,
											"install_port": 22
										}
									]
								}
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "Update cluster"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update cluster data.",
				"summary": "Update cluster.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterUpdateRequest"
							},
							"example": {
								"fwcloud": 2,
								"clusterData": {
									"cluster": 5,
									"name": "Cluster-UPDATED",
									"options": 3
								}
							}
						}
					}
				}
			}
		},
		"/cluster/get": {
			"put": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Get cluster data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 2,
											"fwcloud": 2,
											"name": "Cluster-02",
											"comment": null,
											"created_at": "2019-05-17T11:47:00.000Z",
											"updated_at": "2019-05-17T11:47:00.000Z",
											"created_by": 0,
											"updated_by": 0,
											"nodes": [
												{
													"id": 10,
													"cluster": 2,
													"name": "node1",
													"comment": null,
													"status": 3,
													"install_user": "",
													"install_pass": "",
													"save_user_pass": 1,
													"install_interface": null,
													"install_ipobj": null,
													"fwmaster": 1,
													"install_port": 22,
													"interface_name": null,
													"ip_name": null,
													"ip": null,
													"options": 3
												}
											]
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Cluster access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get cluster data.",
				"summary": "Get cluster data.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 5
							}
						}
					}
				}
			}
		},
		"/cluster/cloud/get": {
			"put": {
				"operationId": "GetByCloud",
				"responses": {
					"200": {
						"description": "Get cloud clusters",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyClusterSummary"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"fwcloud": 2,
												"name": "Cluster-01",
												"comment": null,
												"created_at": "2019-05-17T11:46:57.000Z",
												"updated_at": "2019-05-17T11:46:57.000Z",
												"created_by": 0,
												"updated_by": 0
											},
											{
												"id": 2,
												"fwcloud": 2,
												"name": "Cluster-02",
												"comment": null,
												"created_at": "2019-05-17T11:47:00.000Z",
												"updated_at": "2019-05-17T11:47:00.000Z",
												"created_by": 0,
												"updated_by": 0
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Cluster access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all the cluster data for the indicated fwcloud.",
				"summary": "Get cloud clusters.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetByCloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
		"/cluster/fwtocluster": {
			"put": {
				"operationId": "FirewallToCluster",
				"responses": {
					"200": {
						"description": "Firewall converted to cluster",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterConvertResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"insertId": 6
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Convert a firewall into a new cluster.",
				"summary": "Firewall to cluster.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterFirewallToClusterRequest"
							},
							"example": {
								"fwcloud": 2,
								"firewall": 5,
								"node_id": 391
							}
						}
					}
				}
			}
		},
		"/cluster/clustertofw": {
			"put": {
				"operationId": "ClusterToFirewall",
				"responses": {
					"200": {
						"description": "Cluster converted to firewall",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterConvertResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"insertId": 12
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Convert a cluster into a single firewall.",
				"summary": "Cluster to firewall.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterClusterToFirewallRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 6,
								"node_id": 391
							}
						}
					}
				}
			}
		},
		"/cluster/clone": {
			"put": {
				"operationId": "Clone",
				"responses": {
					"200": {
						"description": "Clone cluster",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterCloneResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"insertId": 7
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Tree node access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Clone an existing cluster.",
				"summary": "Clone cluster.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterCloneRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 5,
								"name": "Cluster-CLONED",
								"comment": "Comment for cloned cluster.",
								"node_id": 391
							}
						}
					}
				}
			}
		},
		"/cluster/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterRestrictionsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"FirewallHasInterfaces": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check if there are restrictions for cluster deletion.",
				"summary": "Restrictions for cluster deletion.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 5
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}